feat(DatasetFiles): refactor table
