+++ src/sections/dataset/dataset-files/files-table/FilesTablePagination.tsx
@@ -1,0 +1,66 @@
+import { Table } from '@tanstack/react-table'
+import { File } from '../../../../files/domain/models/File'
+
+interface FilesTablePaginationProps {
+  table: Table<File>
+}
+
+export function FilesTablePagination({ table }: FilesTablePaginationProps) {
+  return (
+    <div className="flex items-center gap-2">
+      <button
+        className="border rounded p-1"
+        onClick={() => table.setPageIndex(0)}
+        disabled={!table.getCanPreviousPage()}>
+        {'<<'}
+      </button>
+      <button
+        className="border rounded p-1"
+        onClick={() => table.previousPage()}
+        disabled={!table.getCanPreviousPage()}>
+        {'<'}
+      </button>
+      <button
+        className="border rounded p-1"
+        onClick={() => table.nextPage()}
+        disabled={!table.getCanNextPage()}>
+        {'>'}
+      </button>
+      <button
+        className="border rounded p-1"
+        onClick={() => table.setPageIndex(table.getPageCount() - 1)}
+        disabled={!table.getCanNextPage()}>
+        {'>>'}
+      </button>
+      <span className="flex items-center gap-1">
+        <div>Page</div>
+        <strong>
+          {table.getState().pagination.pageIndex + 1} of {table.getPageCount()}
+        </strong>
+      </span>
+      <span className="flex items-center gap-1">
+        | Go to page:
+        <input
+          type="number"
+          defaultValue={table.getState().pagination.pageIndex + 1}
+          onChange={(e) => {
+            const page = e.target.value ? Number(e.target.value) - 1 : 0
+            table.setPageIndex(page)
+          }}
+          className="border p-1 rounded w-16"
+        />
+      </span>
+      <select
+        value={table.getState().pagination.pageSize}
+        onChange={(e) => {
+          table.setPageSize(Number(e.target.value))
+        }}>
+        {[10, 20, 30, 40, 50].map((pageSize) => (
+          <option key={pageSize} value={pageSize}>
+            Show {pageSize}
+          </option>
+        ))}
+      </select>
+    </div>
+  )
+}
