+++ src/sections/dataset/dataset-files/files-table/useFilesTable.tsx
@@ -1,0 +1,30 @@
+import { useState } from 'react'
+import { File } from '../../../../files/domain/models/File'
+import {
+  getCoreRowModel,
+  getFilteredRowModel,
+  getPaginationRowModel,
+  useReactTable
+} from '@tanstack/react-table'
+import { columns } from './FilesTableColumnsDefinition'
+
+export function useFilesTable() {
+  const [data, setFilesTableData] = useState<File[]>(() => [])
+  const [rowSelection, setRowSelection] = useState({})
+
+  const table = useReactTable({
+    data,
+    columns,
+    state: {
+      rowSelection
+    },
+    enableRowSelection: true,
+    onRowSelectionChange: setRowSelection,
+    getCoreRowModel: getCoreRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    debugTable: true
+  })
+
+  return { table, setFilesTableData }
+}
